requirements.txt
django
djangorestframework
django-rq
python-redis-lock